einkaufsliste/requirements.txt
Django==1.9
django-filter==1.1.0
djangorestframework==3.8.2
Markdown==2.6.11
